Change `models.py` file to `models` package for maintainability
